# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ticket Booking System API
  version: 1.0.0
  summary: High-concurrency ticket reservation API with zero-oversell guarantees.
  description: |
    Reservation API for general-admission event tickets, designed to stay correct under
    flash-sale contention.

    ## Booking flow

    1. `POST /events/{eventId}/reservations` places a **hold** (`PENDING`) and decrements
       available inventory inside a single database transaction. The hold has a short TTL
       (3 minutes by default).
    2. `POST /reservations/{reservationId}/confirm` pays for the hold and converts it to
       `CONFIRMED`, creating an order.
    3. `DELETE /reservations/{reservationId}` cancels the hold and returns inventory.
    4. A hold not confirmed before `expiresAt` is released automatically by a background
       worker and becomes `EXPIRED`.

    ## Idempotency

    `POST /events/{eventId}/reservations` **requires** an `Idempotency-Key` header. The key
    is enforced by a UNIQUE constraint in PostgreSQL, so retries are safe across application
    replicas.

    - Retrying with the same key returns the **original** reservation with the original
      status code and the header `Idempotency-Replayed: true`. A second hold is never
      created.
    - Reusing a key with a different request body returns `409 IDEMPOTENCY_KEY_CONFLICT`.
    - If an identical request is still in flight, the retry may return
...

...

    ## Error semantics

...
    the human-readable `detail`.

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /events:
    get:
      tags: [Catalog]
      summary: List events
      description: |
        Paginated catalog listing. Served from cache (60 s TTL) and therefore eventually
        consistent. Ticket counts shown here are advisory.
      operationId: listEvents
      security: []
      parameters:
        - name: status
          in: query
          description: Filter by sale status.
          schema:
            $ref: '#/components/schemas/EventStatus'
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: A page of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPage'
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
  /events/{eventId}:
    get:
      tags: [Catalog]
      summary: Get event details
      operationId: getEvent
      security: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /admin/events:
    post:
      tags: [Admin]
      summary: Create an event and its inventory
      description: |
        Creates the event and its `ticket_inventory` row in one transaction with
        `available = total`, `held = 0`, `sold = 0`.
      operationId: createEvent
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created
          headers:
            Location:
              schema: { type: string, format: uri-reference }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/events/{eventId}:
    patch:
      tags: [Admin]
      summary: Update event metadata
      description: |
        Updates catalog metadata and invalidates the cached catalog entry.

        Total inventory may only be **increased**, never reduced below `held + sold`; the
        database rejects any update that would violate
        `available + held + sold = total` or drive a counter negative.
      operationId: updateEvent
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated and cache invalidated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Update would violate an inventory invariant. Code `INVALID_STATE`.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token from `POST /auth/login`. Claims carry the user id and role
        (`USER` or `ADMIN`). The application is stateless, so any replica accepts any
        valid token.
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      description: Event identifier.
      schema: { type: string, format: uuid }
      example: 7f3a1c0e-9a1b-4f2c-9d3e-1a2b3c4d5e6f
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: |
        Optional correlation id. If absent, the server generates one. It is echoed on the
        response, attached to every log line for the request, and propagated onto emitted
        outbox events.
      schema: { type: string, maxLength: 128 }
...
  responses:
    BadRequest:
      description: Malformed request. Codes `VALIDATION_ERROR`, `INVALID_QUANTITY`, `IDEMPOTENCY_KEY_REQUIRED`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing, malformed, or expired token. Code `UNAUTHORIZED`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted — wrong role, or not the owner of the resource. Code `FORBIDDEN`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource does not exist. Codes `EVENT_NOT_FOUND`, `RESERVATION_NOT_FOUND`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    InternalError:
      description: |
        Unexpected server fault. Code `INTERNAL_ERROR`. Always paired with a logged
        `traceId`. Counts against the availability error budget.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem detail with a stable machine-readable `code`.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type.
        title:
          type: string
          description: Short human-readable summary. Stable per `type`.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation of this occurrence. Never parse this.
        instance:
          type: string
          format: uri-reference
          description: URI of the specific occurrence.
        code:
          $ref: '#/components/schemas/ErrorCode'
        traceId:
          type: string
          description: Correlation id for log lookup.
      additionalProperties: true
    ErrorCode:
      type: string
      description: |
        Stable machine-readable outcome. Clients and load tests branch on this value.

        Business rejections (`SOLD_OUT`, `SALE_NOT_STARTED`, `SALE_ENDED`,
        `USER_LIMIT_EXCEEDED`, `RESERVATION_EXPIRED`, `INVALID_STATE`) mean the system
        worked correctly and refused. They do not consume the availability error budget.
      enum:
        - VALIDATION_ERROR
        - INVALID_QUANTITY
        - IDEMPOTENCY_KEY_REQUIRED
        - IDEMPOTENCY_KEY_CONFLICT
        - IDEMPOTENCY_IN_PROGRESS
        - UNAUTHORIZED
        - FORBIDDEN
        - EVENT_NOT_FOUND
        - RESERVATION_NOT_FOUND
        - SALE_NOT_STARTED
        - SALE_ENDED
        - SOLD_OUT
        - USER_LIMIT_EXCEEDED
        - RESERVATION_EXPIRED
        - INVALID_STATE
        - PAYMENT_DECLINED
        - RATE_LIMITED
        - DEPENDENCY_UNAVAILABLE
        - DEPENDENCY_TIMEOUT
        - INTERNAL_ERROR
...
    EventStatus:
      type: string
      enum: [SCHEDULED, ON_SALE, SOLD_OUT, CLOSED, CANCELLED]
    Event:
      type: object
      required: [id, name, venue, status, saleStartsAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Aurora Live - Opening Night }
        venue: { type: string, example: Riverside Arena }
        startsAt: { type: string, format: date-time }
        saleStartsAt: { type: string, format: date-time }
        saleEndsAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/EventStatus' }
        priceMinor:
          type: integer
          description: Unit price in minor currency units (for example cents).
          example: 4500
        currency: { type: string, minLength: 3, maxLength: 3, example: EUR }
        availability: { $ref: '#/components/schemas/Availability' }
    EventPage:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Event' }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
    Availability:
      type: object
      description: |
        **Advisory snapshot.** May be cached and stale. Never a guarantee that a reservation
        will succeed; only the reserve transaction decides that.
      required: [eventId, available, advisory]
      properties:
        eventId: { type: string, format: uuid }
        total: { type: integer, minimum: 0, example: 100 }
        available: { type: integer, minimum: 0, example: 37 }
        advisory:
          type: boolean
          const: true
          description: Always `true`. A reminder that this value is not authoritative.
        asOf:
          type: string
          format: date-time
          description: When this snapshot was taken.
...
    CreateEventRequest:
      type: object
      required: [name, venue, saleStartsAt, totalTickets, priceMinor, currency]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        venue: { type: string, minLength: 1, maxLength: 200 }
        startsAt: { type: string, format: date-time }
        saleStartsAt: { type: string, format: date-time }
        saleEndsAt: { type: string, format: date-time }
        totalTickets:
          type: integer
          minimum: 1
          description: Initial inventory. Sets available = total, held = 0, sold = 0.
          example: 100
        priceMinor: { type: integer, minimum: 0, example: 4500 }
        currency: { type: string, minLength: 3, maxLength: 3, example: EUR }
      additionalProperties: false
    UpdateEventRequest:
      type: object
      description: |
        Partial update. Inventory may only be increased; the database rejects any change
        that would break `available + held + sold = total` or drive a counter negative.
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        venue: { type: string, minLength: 1, maxLength: 200 }
        startsAt: { type: string, format: date-time }
        saleStartsAt: { type: string, format: date-time }
        saleEndsAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/EventStatus' }
        priceMinor: { type: integer, minimum: 0 }
        addTickets:
          type: integer
          minimum: 1
          description: Increase total and available inventory by this amount.
      additionalProperties: false