# src/main/java/com/ticketsystem/catalog/api/AdminCatalogController.java
package com.ticketsystem.catalog.api;

import com.ticketsystem.catalog.application.AdminCatalogService;
import jakarta.validation.Valid;
import java.net.URI;
import java.util.UUID;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/** ADMIN-only catalog mutations; path authorization is enforced by the security policy. */
@RestController
@RequestMapping("/api/v1/admin/events")
public class AdminCatalogController {

    private final AdminCatalogService adminCatalogService;

    public AdminCatalogController(AdminCatalogService adminCatalogService) {
        this.adminCatalogService = adminCatalogService;
    }

    @PostMapping
    public ResponseEntity<EventResponse> createEvent(
            @Valid @RequestBody CreateEventRequest request) {
        EventResponse created = adminCatalogService.create(request);
        return ResponseEntity.created(URI.create("/api/v1/events/" + created.id())).body(created);
    }

    @PatchMapping("/{eventId}")
    public EventResponse updateEvent(@PathVariable UUID eventId,
                                     @Valid @RequestBody UpdateEventRequest request) {
        return adminCatalogService.update(eventId, request);
    }
}